added variable checks to script

+++ server_build_script.sh
@@ -1,3 +1,17 @@
+# add a check that exits if the variables are not set and exit with an error message
+if [ -z "$JEKYLL_DIR" ]; then
+  echo "JEKYLL_DIR is not set. Exiting."
+  exit 1
+fi
+if [ -z "$NGINX_DIR" ]; then
+  echo "NGINX_DIR is not set. Exiting."
+  exit 1
+fi
+if [ -z "$JEKYLL_BUILDER_IMAGE" ]; then
+  echo "JEKYLL_BUILDER_IMAGE is not set. Exiting."
+  exit 1
+fi
+
 cd $JEKYLL_DIR
 git reset --hard HEAD
 git pull
